+++ src/main/java/fi/stardex/sisu/ui/controllers/dialogs/NewEditTestDialogController.java
@@ -76,7 +76,7 @@ public class NewEditTestDialogController {
     private void createAndSave() {
         InjectorTest injectorTest = new InjectorTest(currentInjectorObtainer.getInjector(), testComboBox.getSelectionModel().getSelectedItem(),
                 Integer.valueOf(rpmTF.getText()), Integer.valueOf(barTF.getText()), Integer.valueOf(adjTimeTF.getText()), Integer.valueOf(measureTimeTF.getText()),
-                Integer.valueOf(freqTF.getText()), Double.valueOf(widthTF.getText()), Double.valueOf(nominalTF.getText()), Double.valueOf(flowRangeTF.getText()));
+                Integer.valueOf(freqTF.getText()), Integer.valueOf(widthTF.getText()), Double.valueOf(nominalTF.getText()), Double.valueOf(flowRangeTF.getText()));
 
         injectorTestRepository.save(injectorTest);
         testListView.getItems().add(injectorTest);
